02_practical-task/requirements.txt
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.1
schedule==1.2.0
aiohttp==3.8.6